+++ src/playground/layout/smart-home/components/app.component.ts
